+++ CHANGELOG.md
@@ -6,20 +6,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+
+## [Unreleased]
+
+### Added
+- type aum_test_function_t
+
 ## [0.10.0](https://github.com/airbus-cyber/aum/releases/tag/0.10.0)
 
 ### Added
 - function `aum_mock_will_skip` configures a mock so that it does nothing when called. Applies only to mock of function with a `void` return type.
+
 
 ## [0.9.2](https://github.com/airbus-cyber/aum/releases/tag/0.9.2)
 
 ### Fixed
 - calls to a mocked function with a `jmp_buf` argument do not fail anymore. However, the `jmp_buf` argument is not stored correctly. And so, for these functions, macros `AUM_ASSERT_WAS_CALLED_WITH` and `AUM_ASSERT_WAS_CALLED_WITH_AT` will not work as expected.
 
+
 ## [0.9.1](https://github.com/airbus-cyber/aum/releases/tag/0.9.1)
 
 ### Fixed
 - calls to a mocked function with a `va_list` argument do not fail anymore. However, the `va_list` argument is not stored correctly. And so, for these functions, macros `AUM_ASSERT_WAS_CALLED_WITH` and `AUM_ASSERT_WAS_CALLED_WITH_AT` will not work as expected.
+
 
 ## [0.9.0](https://github.com/airbus-cyber/aum/releases/tag/0.9.0)