+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.11.1](https://github.com/airbus-cyber/aum/releases/tag/0.11.1)
+
+### Fixed
+- compilation error when option -fno-common is activated (as is the default since gcc 10)
+
+
 ## [0.11.0](https://github.com/airbus-cyber/aum/releases/tag/0.11.0)
 
 ### Added